Edited iRobot Test File
Added ability to define angles, however still working on getting it precise. Also fixed some bugs and errors here and there as well as started to work on getting data from the iRobot. As of current, the GetData() function is not working as it should; will edit to make work later